Don't kill existing processes when connecting
Otherwise you can accidently teardown another process's connnection
when closing the second one you opened. Move this utility into a
separate function `pymmcore_plus.server.try_kill_server` in case it's
necessary for a niche use case or development.
