+++ pymmcore_plus/server/__init__.py
@@ -5,7 +5,15 @@ __all__ = [
     "DEFAULT_URI",
     "pyroCMMCore",
     "serve",
+    "try_kill_server",
 ]
 
 from ._pyrocore import pyroCMMCore
-from ._server import CORE_NAME, DEFAULT_HOST, DEFAULT_PORT, DEFAULT_URI, serve
+from ._server import (
+    CORE_NAME,
+    DEFAULT_HOST,
+    DEFAULT_PORT,
+    DEFAULT_URI,
+    serve,
+    try_kill_server,
+)
